+++ CHANGELOG.md
@@ -8,12 +8,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ### Added
 - Absolute time limit for a fetch job (default: 300 seconds) (@robertcheramy)
 
+### Changed
+- slackdiff: Attempt to join the channel if Errors::NotInChannel is encountered (@varesa)
+
 ### Fixed
 - SSH raises error when closing closed connection. Fixes #3583 (@ytti)
 - Config vars will not fall back to less specific. Fixes #3536 (@ytti)
 - input/scp: make common errors produce a warning, not a crashfile (@robertcheramy)
 - input/scp: implement timeouts. Fixes #3590 (@robertcheramy, @ytti)
 - model/mtrlrfs: add missing prompt (@R3thos)
+- slackdiff: Respect the HTTP proxy configuration while uploading the file. Fixes #3534 (@varesa)
 
 
 ## [0.34.1 - 2025-07-18]